feat: abstract actions, getters, mutations out into their own files
